size_t fixes for string::npos sam got a delay
git-svn-id: https://svn.code.sf.net/p/relacs/code/relacs/trunk@1263 1550813e-0d35-0410-aec7-fc86b090456a

+++ datafile/datatools/bindata.cc
@@ -479,7 +479,7 @@ void readData( DataFile &sf )
     acol[k] = int( acols[k].number( -1.0 ) );
     if ( acol[k] <= 0 )
       amode[k] = 1;
-    unsigned int i = acols[k].rfind( '>' );
+    size_t i = acols[k].rfind( '>' );
     if ( i != string::npos )
       i++;
     else {
